PR #946: Fixing PVS-Studio warning (Contributed by PaulEremeeff)
https://github.com/EpicGames/UnrealEngine/pull/946

I have gone through these changes and verified that whatever behavior change they cause could only be an improvement (generally they just prevent possible access to nullptrs, but there are also some style corrections). A separate changelist is being prepared that includes changes that could unintentionally caues behavior change for the worse.

[CL 2486801 by Dan Oconnor in Main branch]

+++ Engine/Source/Editor/FontEditor/Private/SFontEditorViewport.cpp
@@ -133,7 +133,7 @@ void FFontEditorViewportClient::Draw(FViewport* Viewport, FCanvas* Canvas)
 					RenderInfo->Renderer->GetThumbnailSize(Texture, 1.0f, Width, Height);
 
 					// Don't draw if we are outside of our range
-					if (YPos + Height >= 0 && YPos <= (int32)Viewport->GetSizeXY().Y)
+					if (YPos + (int32)Height >= 0 && YPos <= (int32)Viewport->GetSizeXY().Y)
 					{
 						// If hit testing, draw a tile instead
 						if (Canvas->IsHitTesting())
